Reconstruct the whole structure
