+++ config/example_assignment_cpp_grading_config.json
