--- example-assignment-cpp/example_roster.csv
@@ -1,2 +1,0 @@
-id,last_name,first_name,git_ssh,repo_name
-1,Song,Xinchao,git@github.com:xinchaosong/barreleye-autograder.git,simple-autograder
