--- example-assignment-cpp/git_config.json
@@ -1,6 +1,0 @@
-{
-  "git_config": {
-    "ssh_key_path": "~/.ssh/example_ssh",
-    "roster_path": "example_roster.csv"
-  }
-}
