--- git-scripts/roster.csv
@@ -1,1 +1,0 @@
-last_name,first_name,git_ssh,repo_name
