--- grading-scripts/test_list.csv
@@ -1,1 +1,0 @@
-id,grading_comment,points
