+++ homework/song_xinchao/example-assignment-c/main.c
