+++ scripts/grade.py
@@ -5,32 +5,18 @@
 # Copyright © 2020 Xinchao Song. All rights reserved.
 
 import argparse
-import os
 import json
 import csv
 import subprocess
 
+from scripts import util
+
 
 def load_config():
-    with open("grading_config.json", 'r') as load_f:
+    with open("../config/grading_config.json", 'r') as load_f:
         load_dict = json.load(load_f)
 
     return load_dict['config']
-
-
-def load_csv(csv_path):
-    data_sheet = {}
-
-    if not os.path.exists(csv_path):
-        csv_path = os.path.join(os.path.dirname(__file__), csv_path)
-
-    with open(csv_path, 'r') as f:
-        reader = csv.DictReader(f)
-
-        for row in reader:
-            data_sheet[row['id']] = row
-
-    return data_sheet
 
 
 def save_grades(csv_path, roster):
@@ -175,8 +161,8 @@ def grade(sid, run_test, check_leak):
     configs = load_config()
 
     for i_config in configs.values():
-        roster = load_csv(i_config['roster_path'])
-        test_list = load_csv(i_config['tests_list'])
+        roster = util.load_csv(i_config['roster_path'])
+        test_list = util.load_csv(i_config['tests_list'])
 
         if sid is None:
             for i_student in roster.values():
